+++ apps/purchasing/models/__init__.py
@@ -1,0 +1,18 @@
+"""
+Purchasing app models.
+
+This module imports all models from their individual files to maintain
+backward compatibility and provide a single import point.
+"""
+
+from .purchase_order import PurchaseOrder
+from .purchase_order_line import PurchaseOrderLine
+from .purchase_order_supplier_quote import PurchaseOrderSupplierQuote
+from .stock import Stock
+
+__all__ = [
+    'PurchaseOrder',
+    'PurchaseOrderLine', 
+    'PurchaseOrderSupplierQuote',
+    'Stock',
+]
